requirements.txt
fastapi~=0.95.1
pydantic~=1.10.7
SQLAlchemy~=2.0.12
starlette~=0.26.1
passlib~=1.7.4
python-jose[cryptography]
bcrypt
uvicorn
python-multipart